requirements.txt
PyQt6>=6.5.0
requests>=2.28.0
Pillow>=9.0.0
aiohttp>=3.8.0
aiosqlite>=0.17.0
cryptography>=40.0.0
keyring>=23.0.0
beautifulsoup4>=4.11.0
lxml>=4.8.0
python-dateutil>=2.8.0
psutil>=5.9.0